Migrate link and announcement actions out of admin namespace
