+++ app/controllers/courses_controller.rb
@@ -12,8 +12,8 @@ class CoursesController < ApplicationController
     actions_builder = ActionsBuilder.new(current_holder)
     actions_builder.require(:edit, @course).add_action("Edit Course", :get, edit_admin_course_path(department_id: @course.department))
     actions_builder.require(:edit, @course).add_action("Manage Collaborators", :get, edit_admin_collaborator_group_path(@course.collaborator_group))
-    actions_builder.require(:edit, @course).add_action("Make Announcement", :get, new_admin_announcement_path(course_id: @course.id))
-    actions_builder.require(:edit, @course).add_action("Make Link", :get, new_admin_link_path(course_id: @course.id))
+    actions_builder.require(:edit, @course).add_action("Make Announcement", :get, new_announcement_path(course_id: @course.id))
+    actions_builder.require(:edit, @course).add_action("Make Link", :get, new_link_path(course_id: @course.id))
     actions_builder.require(:destroy, @course).add_action("Destroy Course", :delete, admin_course_path(@course), data: { confirm: 'Are you sure?' })
     @actions = actions_builder.actions
   end
