+++ app/controllers/departments_controller.rb
@@ -19,8 +19,8 @@ class DepartmentsController < ApplicationController
     actions_builder.require(:edit, @department).add_action("Edit Department", :get, edit_admin_department_path(@department))
     actions_builder.require(:edit, @department).add_action("Manage Collaborators", :get, edit_admin_collaborator_group_path(@department.collaborator_group))
     actions_builder.require(:edit, @department).add_action("Make Course", :get, new_admin_department_course_path(department_id: @department.id))
-    actions_builder.require(:edit, @department).add_action("Make Announcement", :get, new_admin_announcement_path(department_id: @department.id))
-    actions_builder.require(:edit, @department).add_action("Make Link", :get, new_admin_link_path(department_id: @department.id))
+    actions_builder.require(:edit, @department).add_action("Make Announcement", :get, new_announcement_path(department_id: @department.id))
+    actions_builder.require(:edit, @department).add_action("Make Link", :get, new_link_path(department_id: @department.id))
     actions_builder.require(:destroy, @department).add_action("Destroy Department", :delete, admin_department_path(@department), data: { confirm: 'Are you sure?' })
     @actions = actions_builder.actions
   end
