+++ app/views/activities/show.haml
@@ -38,7 +38,7 @@
               = content_tag_for(:li, link, class: "list-group-item") do
                 - if current_holder.can?(:edit, link)
                   %i.handle.fa.fa-sort{"aria-hidden" => "true"}
-                  = link_to admin_link_path(link), method: :delete, data: { confirm: 'Are you sure?' } do
+                  = link_to link_path(link), method: :delete, data: { confirm: 'Are you sure?' } do
                     %i.fa.fa-times
                   = link_to new_admin_escalation_request_path(link_id: link) do
                     - if link.escalated?
