+++ app/views/announcements/_form.haml
@@ -1,4 +1,4 @@
-= form_for [:admin, @announcement] do |f|
+= form_for @announcement do |f|
   = hidden_field_tag "#{@announceable.class.to_s.downcase}_id", "#{@announceable.id}"
   .form-group
     = f.label :name
