+++ app/views/announcements/new.haml
