+++ app/views/links/_form.haml
@@ -1,4 +1,4 @@
-= form_for [:admin, @link] do |f|
+= form_for @link do |f|
   = hidden_field_tag "#{@linkable.class.to_s.downcase}_id", "#{@linkable.id}"
   .form-group
     = f.label :name
