+++ app/views/links/new.haml
